requirements.txt
aiogram
python-dotenv
google-genai
sqlalchemy
aiosqlite
asyncpg
fastapi
uvicorn[standard]
httpx